+++ BaseManager.php
@@ -11,14 +11,18 @@ use Yii;
 
 class BaseManager implements ActiveRecordHistoryInterface
 {
+
+    protected $_options;
+
     /**
      * @param integer $type
      * @param \yii\db\ActiveRecord $object
      * @param array $updatedFields
-     * @return void
+     * @param array $options
      */
-    public static function run($type, $object, $updatedFields = [])
+    public static function run($type, $object, $updatedFields = [], $options = [])
     {
+
         $pk = $object->getPrimaryKey(true);
 
         $data = [
@@ -31,7 +35,7 @@ class BaseManager implements ActiveRecordHistoryInterface
         switch ($type) {
             case self::AR_INSERT:
                 $data['field_name'] = array_keys($pk)[0];
-                static::saveField($data);
+                static::saveField($data, $options);
                 break;
             case self::AR_UPDATE:
                 foreach ($updatedFields as $updatedFieldKey => $updatedFieldValue) {
@@ -39,27 +43,27 @@ class BaseManager implements ActiveRecordHistoryInterface
                     $data['old_value'] = $updatedFieldValue;
                     $data['new_value'] = $object->$updatedFieldKey;
 
-                    static::saveField($data);
+                    static::saveField($data, $options);
                 }
                 break;
             case self::AR_DELETE:
                 $data['field_name'] = array_keys($pk)[0];
-                static::saveField($data);
+                static::saveField($data, $options);
                 break;
             case self::AR_UPDATE_PK:
                 $data['field_name'] = array_keys($pk)[0];
                 $data['old_value'] = $object->getOldPrimaryKey();
                 $data['new_value'] = $object->$pk[0];
-                static::saveField($data);
+                static::saveField($data, $options);
                 break;
         }
     }
 
     /**
      * @param array $data
-     * @return void
+     * @param array $options
      */
-    public static function saveField($data)
+    public static function saveField($data, $options = [])
     {
     }